pulumi: Update on pushes, track changes by hashing the project
This change updates the Pulumi workflow to run `pulumi up` on pushes to `main`
and `pulumi preview` on other branches. It also changes the `ZippedRustBinary`
resource to track changes by hashing the project directory instead of the last
modified file. This is because `git` doesn't track the last modified time of
files, so it's not a reliable way to determine if the project has changed.