build/oidc: Add prod stack, build in cache
The build is slow, and we're currently iterating on the OIDC setup. This
slowness means that iterating is taking a long time. Since we build the Rust
binary in CI anyway, we can share the results between the build test and the
deploy job.

Additionally, we're adding a prod stack which will be used to deploy the
application to a production environment from CI. This should let us test and
manually apply locally and then deploy to production from CI.

Finally, update one or two policies to add permissions we've already seen we
need.