CI: Don't run build/deploy jobs concurrently on pushes
Currently each job tries to run concurrently, because we consider each commit
its own unit. But when multiple pushes happen in quick succession, the jobs
will run concurrently and may interfere with each other. This is especially
problematic for the deploy job, which can only be run by one job at a time since
it - correctly - takes a lock.

Fix this by still allowing concurrent builds for PRs, but serialising the
build/deploy jobs for pushes.