build: Fix caching of resulting binaries
The `Swatinem/rust-cache` action cleans the workspace before saving the cache.
That's fine for building, but we also want to cache the built binaries between
jobs too. For two reasons:

1. To avoid building it on subsequent pushes when the source files haven't
   changed.
2. We build once in either the `build` or `oidc` workflows (whichever comes
   first) and then use the same binary in the other workflow.

For either of these the actual binaries themselves need to be restored. Switch
to using `actions/cache/restore` and `actions/cache/save` directly, which allows
us to control when the saving/restoring happens. We need to do this, because
`Swatinem/rust-cache` will run its `post` step to save the cache _before_
`actions/cache` would save its cache. And that erases our built artifacts, so
when `actions/cache`'s `post` step gets to it, the files are gone.

Now then. We were actually caching the depdencies twice before, once using
`actions/cache` and once via `actions-rust-lang/setup-rust-toolchain`'s usage of
`Swatinem/rust-cache`. That's fixed here to keep two caches: our binaries, and
the Rust dependencies. There's no point in restoring the binaries if there's a
non-exact match, since we will rebuild anyway, so drop `restore-keys`.

But then we discover that this hid a bug. We we want to skip rebuilding in the
cache of a cache hit - that's half the point of using the cache - but we _do_
want to run the tests anyway (now that I think about it: if nothing changed, why
do we? don't want to think about that now...). So we let the `build` job ask for
the cache to be restored.

One more thing. We should rebuild in the case the toolchain changes, not only
our own source files. The `if` conditions are altered slightly so that the build
runs if _either_ cache has a miss.