CI: Fix Actions concurrency with pushes
We use the built artifacts for arm64 in two jobs. To save effort, we have set up
a concurrency group so that the build happens once and the second fetches it
from the cache.

This worked okay in PRs, but for pushes a problem emerged. The behaviour we want
is that each push to `main` gets a deployment in turn, or alternatively gets to
complete before the latest one starts. If there are multiple in
a short space, what we're actually seeing is in-progress jobs being cancelled.
This is because GitHub Actions only allows one queued job per concurrency group
ever.

It happens like this:

  - Push 1 starts, build job starts, deployment job queued
  - Push 2 arrives and _cancels_ both the in-progress and queued jobs from push
    1

Here's an attempt at solving this in two ways. First, we rework the concurrency
groups. The build jobs (in two different workflows btw) will share a group as
before with `cancel-in-progress: false`. Next we'll add a workflow-level
concurrency groups so that PR builds cancel earlier ones (to not waste resources
on building superseded commits) but pushes don't so that the below queue is
used.

And finally we add [`turnstyle`][turnstyle] to both workflows. This is
where the workaround for GitHub's limitation comes in. It will ensure that only
one job runs at a time per-branch for the workflow it's running in, by waiting
in a loop (up to a 10 minute timeout) until it's the only one running.

[turnstyle]: https://github.com/marketplace/actions/action-turnstyle