Switch to pnpm
We're getting an obscure error about [not being able to find a module][error].

```
Error: Cannot find module '@jridgewell/trace-mapping'
```

If you look in the `yarn info` output, this module _is_ present. But in "print
dir contents" it's not there. I've given up debugging it, since it seems
impossible to get `yarn` to be more verbose. Switching to `pnpm` works. Let's do
that.

[error]: https://github.com/iainlane/rssfilter/actions/runs/10026623776/job/27711155594