ci(pulumi): wait on turnstile job
We're running `turnstile` but not awaiting it, so the `pulumi` job is running
before the `turnstile` job has finished. This is causing the `pulumi` job to
fail sometimes, when concurrent `pulumi` jobs are trying to access the same
state. Move the step into the `pulumi` job to fix this. We can't do this with
two jobs, as `needs` can't be conditional, and we only wait for push events.