feat: drop Pulumi dynamic provider and build ourselves, use rust tls
This breaks every time Pulumi is updated. It's something to do with
mismatching versions of Pulumi packages. Building ourselves fixes this.
Add a `build-rust-lambda.ts` script to build the lambda into a zip file.
It's then found by Pulumi and uploaded.

Switch `reqwest` to use `rustls` instead of `native-tls` too. By default
it uses openssl and this is a pain to cross compile.