ci: use rust-cache for target
We're rebuilding unnecessarily. This is making us fail in the "Make zip
file" step, because the cross toolchain isn't installed there. It
shouldn't have to rebuild, so the toolchain ought not to be necessary.
Here we try to fix this by letting `Swatinem/rust-cache` handle the
cache, hoping it'll get everything.