chore: update Pulumi to use Cloudflare for DNS and TLS
This doesn't work - the `Host` header is wrong when it's proxied through
Cloudflare, and so the Lambda doesn't get served.

The next commit will switch to using the Cloudflare Workers. This is an
intermediate state, just to save the Pulumi changes I made previously,
in case they turn out to be useful later.