feat: port to Cloudflare Workers and wasm32
I noticed that I was getting charged for ACM certs on AWS. In response,
I tried to move DNS over to Cloudflare Workers. That didn't work, of
course. What happens when Workers proxies to Lambda is that the `Host`
header is wrong, because the Lambda doesn't serve the custom domain any
more.

In response to that, I decided to see how hard it would be to use
`workers-rs` and run on Cloudflare Workers. Well... quite hard. I didn't
know anything about WASM before really, apart from that it's a way to
run Rust (or any compiled) code on the web.

"On the web" is the key bit here. It's quite a different platform and
you can't just pretend you're writing native code. I've ended up
conditionally compiling two versions here: a native one using `reqwest`,
and a web one using `fetch`.

This meant all the internal types had to be refactored to keep a
sensible abstraction. So now we have a translation layer between the web
`Request` and `Response` types and `reqwest`'s.

I've not been able to make OTLP work in WASM yet. The libraries pull in
`tokio` and this ends up depending on things which doesn't work on WASM.
I had wanted to make it start sending directly to Grafana Cloud. But
that's not there now. We do have OTel working, logging to stdout,
though.

This also decommissions the AWS resources, to hopefully cut that bill
off.