test: replace httpbin.org with fake HTTP client
Some of the WASM were depending on httpbin.org which is unreliable and
was making them flake out. Replace this with a fake HTTP client that's
provided to the RSS client using dependency injection.

This was only affecting the WASM environment, but we've updated the
tests for WASM and non-WASM for consistency.